// src/types.ts
export type coinElement = {
  id: number;
  name: string;
  price: string;
  price30m: string;
  price1h: string;
  price6h: string;
  price12h: string;
  price24h: string;
};

export type coinElementArray = {
  id: number;
  name: string;
  price: string;
  price30m: string;
  price1h: string;
  price6h: string;
  price12h: string;
  price24h: string;
}[];